Fix UDP tests
