+++ core/src/test/java/io/questdb/test/cutlass/line/udp/LineUdpPartitionReadOnlyTest.java
@@ -41,12 +41,25 @@ import io.questdb.test.cairo.TableModel;
 import io.questdb.test.cutlass.line.AbstractLinePartitionReadOnlyTest;
 import io.questdb.test.tools.TestUtils;
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 
 import static io.questdb.cairo.TableUtils.createTable;
 import static io.questdb.test.tools.TestUtils.*;
 
 public class LineUdpPartitionReadOnlyTest extends AbstractLinePartitionReadOnlyTest {
+
+    @Before
+    public void setUp() {
+        super.setUp();
+        TestUtils.unchecked(() -> createDummyConfiguration(
+                        "cairo.max.uncommitted.rows=500",
+                        "cairo.commit.lag=2000",
+                        "cairo.o3.max.lag=2000",
+                        "line.udp.enabled=true"
+                )
+        );
+    }
 
     @Test
     public void testActivePartitionReadOnlyAndNoO3UDP() throws Exception {
